requirements.txt
tensorflow-gpu == 2.5.0
tensorflow_addons >= 0.19.0
scipy >= 1.7.3

